Naive asyncification
